+++ test/react-web/client/graphql/queries.tsx
@@ -1127,4 +1127,59 @@ describe('queries', () => {
       </ProviderMock>);
   });
 
+  it('exposes updateQuery as part of the props api', (done) => {
+    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
+    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
+    const networkInterface = mockNetworkInterface({ request: { query }, result: { data } });
+    const client = new ApolloClient({ networkInterface });
+
+    @graphql(query)
+    class Container extends React.Component<any, any> {
+      componentWillReceiveProps({ data }) { // tslint:disable-line
+        expect(data.updateQuery).to.be.exist;
+        expect(data.updateQuery).to.be.instanceof(Function);
+        expect(data.updateQuery).to.not.throw;
+        done();
+      }
+      render() {
+        return null;
+      }
+    };
+
+    mount(<ProviderMock client={client}><Container /></ProviderMock>);
+  });
+
+  it('allows updating query results after query has finished', (done) => {
+    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
+    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
+    const data2 = { allPeople: { people: [ { name: 'Leia Skywalker' } ] } };
+    const networkInterface = mockNetworkInterface(
+      { request: { query }, result: { data } },
+      { request: { query }, result: { data: data2 } }
+    );
+    const client = new ApolloClient({ networkInterface });
+
+    let isUpdated;
+    @graphql(query)
+    class Container extends React.Component<any, any> {
+      componentWillReceiveProps(props) {
+        if (isUpdated) {
+          expect(props.data.allPeople).to.deep.equal(data2.allPeople);
+          done();
+          return;
+        } else {
+          isUpdated = true;
+          props.data.updateQuery((prev) => {
+            return data2;
+          });
+        }
+      }
+      render() {
+        return null;
+      }
+    };
+
+    mount(<ProviderMock client={client}><Container /></ProviderMock>);
+  });
+
 });